fix : 주문량 상위 20% 책 데이터 redis에 저장

+++ src/main/java/com/fastcampus/book_bot/dto/order/OrdersDTO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Builder
 public class OrdersDTO {
 
-    private Long bookId;
+    private Integer bookId;
     private String bookName;
     private String bookAuthor;
     private String bookPublisher;
